requirements.txt
fastapi
uvicorn
jinja2
python-multipart
tensorflow
librosa
numpy
pillow
scikit-learn
matplotlib
